Preserve prototype when setting an object property
Object.assign does not copy properties in the prototype chain. In order to preserve them, we create a new object using
the prototype via Object.create. If the user is trying to set a property that was in the prototype, it will still be set
via Object.assign and will be found first.